% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadData.R
\name{add.in.cn.as.single.snv}
\alias{add.in.cn.as.single.snv}
\title{Function that adds copy number as a single SNV into a data set}
\usage{
add.in.cn.as.single.snv(dataset, cndata, cellularity, add.conflicts = T)
}
\description{
Function that adds copy number as a single SNV into a data set
}
