% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadData.R
\name{add.mutphasing}
\alias{add.mutphasing}
\title{Add mutation phasing information to a dataset}
\usage{
add.mutphasing(dataset, mutphasing, add.conflicts = F)
}
\arguments{
\item{dataset}{A data set object}

\item{mutphasing}{data.frame with phasing data}

\item{add.conflicts}{Supply TRUE when a mutation-to-mutation confict matrix is to be built (Default: FALSE)}
}
\value{
a dataset with field mutphasing and potentially field conflict.array added
}
\description{
Add mutation phasing information to a dataset
}
