% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirichletProcessClustering.R
\name{add_removed_snvs}
\alias{add_removed_snvs}
\title{Add removed mutations back into the assignment table. SNVs will be assigned to the cluster of its most similar not-removed SNV}
\usage{
add_removed_snvs(dataset, snv_assignment_table)
}
\arguments{
\item{dataset}{A dataset object}

\item{snv_assignment_table}{Data frame with the mutation assignments}
}
\value{
The snv_assignment_table with the removed mutations added into the position they were originally
}
\description{
Add removed mutations back into the assignment table. SNVs will be assigned to the cluster of its most similar not-removed SNV
}
\author{
sd11
}
