% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadData.R
\name{append.dataset}
\alias{append.dataset}
\title{Convenience function to append two datasets}
\usage{
append.dataset(a, b)
}
\arguments{
\item{a}{dataset a}

\item{b}{dataset b}
}
\value{
a dataset in which a and b are appended
}
\description{
Convenience function to append two datasets
}
