% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirichletProcessClustering.R
\name{assign_cnas_to_clusters}
\alias{assign_cnas_to_clusters}
\title{Assign CNA events to clusters using their pseudoSNV representation}
\usage{
assign_cnas_to_clusters(cndata, snv_assignment_table)
}
\arguments{
\item{cndata}{Data frame with the CNA data}

\item{snv_assignment_table}{Data frame with the mutation assignments, with chromosome and position-start/end expected as first three columns}
}
\value{
The cndata object with extra column cluster_assignment
}
\description{
Assign CNA events to clusters using their pseudoSNV representation
}
\author{
sd11
}
