% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{build_coassignment_prob_matrix_densities}
\alias{build_coassignment_prob_matrix_densities}
\title{Build a coassignment probability matrix using SNV specific densities over
the assigned CCFs during MCMC}
\usage{
build_coassignment_prob_matrix_densities(S.i, pi.h, no.iters.burn.in)
}
\arguments{
\item{S.i}{The mutation assignments to cluster ids}

\item{pi.h}{The cluster CCFs}

\item{no.iters.burn.in}{The number of iterations to discard as burn in}
}
\value{
Square matrix with a row and column for every mutation, each cell contains the probability of the pair of mutations belonging to the same cluster
}
\description{
Build a coassignment probability matrix using SNV specific densities over
the assigned CCFs during MCMC
}
\author{
dw9, sd11
}
