% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadData.R
\name{create_pseudo_snv}
\alias{create_pseudo_snv}
\title{Create a number of pseudo SNVs to represent a CNA and add them to the dataset}
\usage{
create_pseudo_snv(
  cndata.i,
  num_muts,
  N,
  conf,
  cellularity,
  dataset,
  conflicting.events.only
)
}
\description{
Create a number of pseudo SNVs to represent a CNA and add them to the dataset
}
