% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triplet_functions.R
\name{determine_num_triplets}
\alias{determine_num_triplets}
\title{Determines the total number of triplets from the given number of samples}
\usage{
determine_num_triplets(num_samples)
}
\arguments{
\item{num_samples}{Integer total number of samples}
}
\value{
Integer total of triplets
}
\description{
Determines the total number of triplets from the given number of samples
}
\author{
Naser Ansari-Pour (BDI, Oxford)
}
