% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirichletProcessClustering.R
\name{DirichletProcessClustering}
\alias{DirichletProcessClustering}
\title{Main function to run subclonal reconstruction}
\usage{
DirichletProcessClustering(
  mutCount,
  WTCount,
  totalCopyNumber,
  copyNumberAdjustment,
  mutation.copy.number,
  cellularity,
  output_folder,
  no.iters,
  no.iters.burn.in,
  subsamplesrun,
  samplename,
  conc_param,
  cluster_conc,
  mut.assignment.type,
  most.similar.mut,
  mutationTypes,
  max.considered.clusters
)
}
\arguments{
\item{mutCount}{Matrix with readcounts of the mutated allele}

\item{WTCount}{Matrix with readcounts of the wild-type allele}

\item{totalCopyNumber}{Matrix with total copynumber at each mutation locus}

\item{copyNumberAdjustment}{Matrix with multiplicity values}

\item{mutation.copy.number}{Matrix with mutation copy number values}

\item{cellularity}{Vector with sample purities}

\item{output_folder}{Directory where to write output}

\item{no.iters}{The number of iterations to run the MCMC chain for}

\item{no.iters.burn.in}{Number of iterations to discard as burn in}

\item{subsamplesrun}{Samplenames of individual samples for this donor}

\item{samplename}{Donor name, used in plots and to name output files}

\item{conc_param}{Hyperparameter setting that affects the sampling of the alpha stick-breaking parameter}

\item{cluster_conc}{Legacy parameter, no longer used}

\item{mut.assignment.type}{Type of mutation assignment to be used}

\item{most.similar.mut}{Vector with most similar mutation for mutations removed during sampling (if any)}

\item{mutationTypes}{Vector with mutation types, used for plotting}

\item{max.considered.clusters}{Maximum number of clusters to consider}
}
\description{
Will perform clustering using the given data. The method
decides automatically whether the 1D or nD method is run based on the number of samples given at the input.
The number of samples is determined through the number of columns of the input.
}
\author{
sd11
}
