% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleMutations.R
\name{do_uniform_sampling}
\alias{do_uniform_sampling}
\title{Uniformly sample mutations}
\usage{
do_uniform_sampling(avail_for_sampling, num_muts_sample)
}
\arguments{
\item{num_muts_sample}{The number of mutations to sample}

\item{num_muts}{The total number of mutations in the dataset}
}
\value{
A vector with indices of mutations to select
}
\description{
Uniformly sample mutations
}
\author{
sd11
}
