% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirichletProcessClustering.R
\name{flatten_3d_to_2d}
\alias{flatten_3d_to_2d}
\title{Helper function that flattens a 3D array into a 2D one}
\usage{
flatten_3d_to_2d(data, col_names)
}
\arguments{
\item{data}{The data to be flattened}

\item{col_names}{The names of the columns in the output}
}
\value{
A data.frame with the third column annotated as the first column
}
\description{
Helper function that flattens a 3D array into a 2D one
}
\author{
sd11
}
