% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadData.R
\name{get.conflicting.indices}
\alias{get.conflicting.indices}
\title{Returns a list of subclonal SNV indices that are conflicting with a CNA event. This code 
classifies a subclonal SNV has a CCF of < 0.9. This function returns a vector consisting of T/F values.
TODO: Determine whether an SNV is likely to be subclonal in a less arbitrary way.}
\usage{
get.conflicting.indices(dataset, cndata)
}
\description{
Returns a list of subclonal SNV indices that are conflicting with a CNA event. This code 
classifies a subclonal SNV has a CCF of < 0.9. This function returns a vector consisting of T/F values.
TODO: Determine whether an SNV is likely to be subclonal in a less arbitrary way.
}
