% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triplet_functions.R
\name{get.subsample.indices}
\alias{get.subsample.indices}
\title{Determines the indices of the samples to be used for a certain triplet}
\usage{
get.subsample.indices(choose.index, choose.number, choose.from)
}
\arguments{
\item{choose.index}{Of all possible combinations of samples, run this combination}

\item{choose.number}{The number within the total triplets to be picked}

\item{choose.from}{The total number of triplets}
}
\value{
Vector of integers
}
\description{
Determines the indices of the samples to be used for a certain triplet
}
\author{
Naser Ansari-Pour (BDI, Oxford)
}
