% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirichletProcessClustering.R
\name{get_cnas_cluster_probs}
\alias{get_cnas_cluster_probs}
\title{Use the Pseudo-SNV probabilities to obtain a probability of each CNA of each cluster}
\usage{
get_cnas_cluster_probs(cndata, snv_assignment_likelihoods, cluster_colnames)
}
\arguments{
\item{cndata}{The copy number data data.frame}

\item{snv_assignment_likelihoods}{Probabilities of the pseudo-SNVs}

\item{cluster_colnames}{The colnames of the output bedfile to be used to select the assignment columns}
}
\value{
A data.frame with chr,start,end,probs_per_cluster
}
\description{
Use the Pseudo-SNV probabilities to obtain a probability of each CNA of each cluster
}
\author{
sd11
}
