% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleMutations.R
\name{get_data_avail_for_sampling}
\alias{get_data_avail_for_sampling}
\title{Function that returns the indices of data that is available for sampling given
that one can choose to only sample SNVs, sample both SNVs and CNAs or remove SNVs
alltogether}
\usage{
get_data_avail_for_sampling(dataset, sample.snvs.only, remove.snvs)
}
\arguments{
\item{dataset}{A dataset object}

\item{sample.snvs.only}{Boolean that if set to TRUE samples SNVs only. CNA indices should then lateron be added manually if they need to stay in the dataset}

\item{remove.snvs}{A boolean that if set to TRUE removes the SNVs from the dataset and returns just the CNA indices}
}
\value{
A list of indices representing the data that can be sampled under the given restrictions
}
\description{
Function that returns the indices of data that is available for sampling given
that one can choose to only sample SNVs, sample both SNVs and CNAs or remove SNVs
alltogether
}
\author{
sd11
}
