% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_snv_assignment_ccfs}
\alias{get_snv_assignment_ccfs}
\title{Fetch the CCFs of the clusters to which SNVs have been assigned during MCMC into a big table}
\usage{
get_snv_assignment_ccfs(
  pi.h,
  S.i,
  no.muts,
  no.timepoints,
  no.iters,
  no.iters.burn.in
)
}
\arguments{
\item{pi.h}{The cluster CCFs}

\item{S.i}{The mutation assignments to cluster ids}

\item{no.muts}{Number of total SNVs}

\item{no.iters}{The total number of iterations}

\item{no.iters.burn.in}{The total number of iterations to use as burnin}
}
\value{
A matrix with a column for each SNV and a row for each to consider iteration with the cell containing the CCF
}
\description{
Fetch the CCFs of the clusters to which SNVs have been assigned during MCMC into a big table
}
\author{
sd11
Get assignment ccfs for each snv
}
