% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignMutations.R
\name{getLocalOptima}
\alias{getLocalOptima}
\title{Function that fetches the local optima from a density function call output}
\usage{
getLocalOptima(cluster_density, hypercube.size = 20)
}
\arguments{
\item{cluster_density}{Density estimate of where clusters are located}

\item{hypercube.size}{Stepsize to use when stepping through the density looking for optima}
}
\value{
A list containing two fields: localOptima, the location of a peak and peak.indices, the index of the peak within a hypercube
}
\description{
Function that fetches the local optima from a density function call output
}
\author{
sd11, dw9
}
