% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadData.R
\name{load.cn.data}
\alias{load.cn.data}
\title{Load the CN input for a single sample (for now)
This expects a cnDP input file}
\usage{
load.cn.data(infile)
}
\description{
Load the CN input for a single sample (for now)
This expects a cnDP input file
}
