% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadData.R
\name{load.data}
\alias{load.data}
\title{This function loads a series of data tables or VCFs into a dataset object
which is in essense a list of tables, one for each type of information}
\usage{
load.data(
  list_of_data_files,
  cellularity,
  Chromosome,
  position,
  WT.count,
  mut.count,
  subclonal.CN,
  no.chrs.bearing.mut,
  mutation.copy.number,
  subclonal.fraction,
  phase = NULL,
  is.male = T,
  is.vcf = F,
  ref.genome.version = "hg19",
  supported_chroms = c(1:22)
)
}
\arguments{
\item{list_of_data_files}{A list of data file names to be read in}

\item{cellularity}{A vector containing the cellularity estimates for each sample}

\item{Chromosome}{String that contains the name of the column with chromosome denomination for each mutation}

\item{position}{String that contains the name of the column with position denomination for each mutation}

\item{WT.count}{String the colname of the column in the data files that contains the number of wild type reads}

\item{mut.count}{String, the colname of the column in the data files that contains the number of mutant reads}

\item{subclonal.CN}{String, the name of the total copynumber column}

\item{no.chrs.bearing.mut}{String, the column that contains the number of chromosomes not bearing any mutations}

\item{mutation.copy.number}{String, colname with mutation copy number estimates}

\item{subclonal.fraction}{String, colname of the column that contains the subclonal fraction (or CCF) estimate for each mutation}

\item{phase}{String, colname of the copy number to mutation phasing info column}

\item{is.male}{Optional boolean that represents the sex, set to TRUE if male, FALSE if female. This information is used to decide whether to include X chromosome mutations}

\item{is.vcf}{Optional boolean parameter whether the files to be read in are in VCF format}

\item{ref.genome.version}{Optional string that represents the reference genome, required when reading in VCF files}
}
\value{
A list of tables, one for each type of information
}
\description{
This function loads a series of data tables or VCFs into a dataset object
which is in essense a list of tables, one for each type of information
}
\author{
sd11
}
