% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadData.R
\name{load.indel.data}
\alias{load.indel.data}
\title{Load the indel input
This expects a DP indel input file}
\usage{
load.indel.data(infiles)
}
\description{
Load the indel input
This expects a DP indel input file
}
