% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirichletProcessClustering.R
\name{make_advanced_params}
\alias{make_advanced_params}
\title{Helper function to package advanced parameters - most of these will almost never need to be changed}
\usage{
make_advanced_params(
  seed,
  conc_param = 0.01,
  cluster_conc = 5,
  max.considered.clusters = 20
)
}
\arguments{
\item{seed}{The seed to use}

\item{conc_param}{Hyperparameter setting that affects the sampling of the alpha stick-breaking parameter}

\item{cluster_conc}{Legacy parameter, no longer used}

\item{max.considered.clusters}{The maximum number of clusters to be considered}
}
\value{
A list containing these components
}
\description{
Helper function to package advanced parameters - most of these will almost never need to be changed
}
\author{
sd11
}
