% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirichletProcessClustering.R
\name{make_cna_params}
\alias{make_cna_params}
\title{Helper function to package CNA parameters - to be implemented}
\usage{
make_cna_params()
}
\description{
Helper function to package CNA parameters - to be implemented
}
