% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirichletProcessClustering.R
\name{make_run_params}
\alias{make_run_params}
\title{Helper function to package run parameters}
\usage{
make_run_params(no.iters, no.iters.burn.in, mut.assignment.type,
  num_muts_sample, is.male, min_muts_cluster = NULL,
  min_frac_muts_cluster = 0.01, species = "human",
  assign_sampled_muts = TRUE, supported_chroms = NULL,
  keep_temp_files = TRUE, generate_cluster_ordering = FALSE,
  memory_limit_gb = NA_real_, num_threads = NA_integer_)
}
\arguments{
\item{no.iters}{The number of iterations that the MCMC chain should be run for}

\item{no.iters.burn.in}{The number of iterations that should be discarded as burn in of the MCMC chain}

\item{mut.assignment.type}{Mutation assignment type option}

\item{num_muts_sample}{The number of mutations from which to start downsampling}

\item{is.male}{Boolean set to TRUE when the donor is male, female otherwise}

\item{min_muts_cluster}{The minimum number of mutations required for a cluster to be kept in the final output (Default: NULL)}

\item{min_frac_muts_cluster}{The minimum fraction of mutations required for a cluster to be kept in the final output (Default: 0.01)}

\item{species}{Species (Default: Human)}

\item{assign_sampled_muts}{A boolean whether to assign mutations that have not been used for clustering due to downsampling (Default: TRUE)}

\item{supported_chroms}{Vector with chromosome names from which mutations can be used (Default: NULL)}

\item{keep_temp_files}{Set to TRUE to keep temporary files (Default: TRUE)}

\item{generate_cluster_ordering}{Set to TRUE to generate possible phylogenetic relationships between clusters (Default: FALSE)}

\item{memory_limit_gb}{Optional hard memory guard rail in GB. DPClust runs if feasible under this limit, otherwise errors before starting.}

\item{num_threads}{Number of CPU threads for the Gibbs C++ core (OpenMP). If NA, use runtime default.}
}
\value{
A list containing these components
}
\description{
Helper function to package run parameters
}
\author{
sd11
}
