% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triplet_functions.R
\name{make_triplet_params}
\alias{make_triplet_params}
\title{Helper function to package triplet pipeline run parameters}
\usage{
make_triplet_params(
  no.iters = 1000,
  no.iters.burn.in = NULL,
  density.smooth = 0.01,
  keep_md_out = TRUE
)
}
\arguments{
\item{no.iters}{The number of iterations that the MCMC chain should be run for (Default: 1000). Used by steps "core" and "density", not used by step "combine". Must be identical between steps "core" and "density"}

\item{no.iters.burn.in}{The number of iterations that should be discarded as burn in of the MCMC chain (Default: ceiling(no.iters/5)). Used by steps "core" and "density", not used by step "combine". Must be identical between steps "core" and "density"}

\item{density.smooth}{Smoothing parameter used during multi-dimensional density estimation (Default: 0.01) - internal tuning parameter, not intended to be changed by regular users. Must be identical between steps "density" and "combine"}

\item{keep_md_out}{Boolean, set to FALSE to remove the "md_out" working directory (the per-triplet intermediate files written by step "density") once step "combine" has copied its final result files out to outdir (Default: TRUE). Only used by step "combine"}
}
\value{
A list containing these components
}
\description{
Helper function to package triplet pipeline run parameters
}
\author{
sd11
}
