% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignMutations.R
\name{mutation_assignment_em}
\alias{mutation_assignment_em}
\title{Assign mutations for multi-dimensional clustering by adding clusters until the likelihood no longer improves}
\usage{
mutation_assignment_em(
  GS.data,
  mutCount,
  WTCount,
  subclonal.fraction,
  node.assignments,
  opts
)
}
\arguments{
\item{GS.data}{Output from the MCMC chain}

\item{mutCount}{Matrix containing counts of the mutated allele}

\item{WTCount}{Matrix containing counts of the wild type allele}

\item{subclonal.fraction}{Matrix with cancer cell fraction estimates for each mutation in each sample}

\item{node.assignments}{Matrix with assignments of mutations to clusters during MCMC (i.e. GS.data$S.i)}

\item{opts}{List with various parameters, including samplename and number of iterations}
}
\value{
Standardised clustering output, including mutation assignments, likelihoods and cluster positions
}
\description{
Assign mutations for multi-dimensional clustering by adding clusters until the likelihood no longer improves
}
\author{
dw9, sd11
}
