% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDensities.R
\name{plot1D}
\alias{plot1D}
\title{Original plotting function: Creates the 1D density plot in either allele fraction, mutation copy number or fraction of tumour cells space.}
\usage{
plot1D(
  density,
  polygon.data,
  pngFile = NA,
  density.from = 0,
  x.max = NA,
  y.max = NA,
  y = NULL,
  N = NULL,
  mutationCopyNumber = NULL,
  no.chrs.bearing.mut = NULL,
  samplename = "",
  CALR = numeric(0),
  cluster.locations = NULL,
  mutation.assignments = NULL,
  mutationTypes = NULL
)
}
\arguments{
\item{density}{Is a two column data frame. First column x-axis density line, second column y-axis density line.}

\item{polygon.data}{Contains the coordinates of the 95\% confidence interval around the density line.}

\item{pngFile}{Where to save the figure (Default: NA)}

\item{density.from}{Startpoint from where density is drawn (Default: 0)}

\item{x.max}{Maximum values on x axis in the plot (Default: NA, determine from data)}

\item{y.max}{Maximum values on y axis in the plot (Default: NA, determine from data)}

\item{y}{MutCount data, only used when mutationCopyNumber is not supplied (Default: NULL)}

\item{N}{Total count (mutCount+WTCount), only used when mutationCopyNumber is not supplied (Default: NULL)}

\item{mutationCopyNumber}{Mutation copy number data, when not supplied y and N are required (Default: NULL)}

\item{no.chrs.bearing.mut}{Copynumber (multiplicity) adjustment per mutation, when not supplied the plot will either contain mutation copy number or allele frequencies (Default: NULL)}

\item{samplename}{Name of the sample under analysis (Default: empty string)}

\item{cluster.locations}{Locations in the space to be plotted where clusters reside (Default: NULL)}

\item{mutation.assignments}{Mutation assignments (Default: NULL)}

\item{mutationTypes}{Type of each mutation, different mutation types can be plot with different colours (Default: NULL)}

\item{CARL}{Legacy annotation parameter, no longer used (Default: empty numeric)}
}
\description{
Original plotting function: Creates the 1D density plot in either allele fraction, mutation copy number or fraction of tumour cells space.
}
\author{
dw9, sd11
}
