% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDensities.R
\name{plotnD}
\alias{plotnD}
\title{Create a 2D plot for multi-dimensional clustering}
\usage{
plotnD(
  xvals,
  yvals,
  zvals,
  subclonal.fraction_x,
  subclonal.fraction_y,
  pngFile,
  samplename_x,
  samplename_y,
  max.plotted.value = NA,
  cluster.locations = NULL,
  plot_mutations = F
)
}
\arguments{
\item{xvals}{Density coordinates of the x-axis}

\item{yvals}{Density coordinates of the y-axis}

\item{zvals}{The density itself}

\item{subclonal.fraction_x}{Fraction of cells estimate for each mutation in sample on x-axis}

\item{subclonal.fraction_y}{Fraction of cells estimate for each mutation in sample on y-axis}

\item{pngFile}{Filename to which the figure should be pushed}

\item{samplename_x}{Samplename to assign as x-axis label}

\item{samplename_y}{Samplename to assign as y-axis label}

\item{max.plotted.value}{The maximum value to plot on both x- and y-axis (Default: NA)}

\item{cluster.locations}{Locations of clusters to be overlayed (Default: NULL)}

\item{plot.mutations}{Boolean, supply TRUE when mutations should be overlayed on top of the density (Default: FALSE)}
}
\description{
Create a 2D plot for multi-dimensional clustering
}
\author{
dw9
}
