% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{read_gsdata_object}
\alias{read_gsdata_object}
\title{Helper function to read in the DP algorithm files}
\usage{
read_gsdata_object(indir, samplename, no.iters, conc_param, cluster_conc)
}
\arguments{
\item{indir}{The directory in which the files were stored}
}
\value{
A GS.data object with fields S.i, V.h, pi.h and alpha
}
\description{
Helper function to read in the DP algorithm files
}
\author{
sd11
}
