% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadData.R
\name{remove_mutations}
\alias{remove_mutations}
\title{Helper function to remove a set of mutations from a dataset by their index}
\usage{
remove_mutations(dataset, mutation_index)
}
\arguments{
\item{dataset}{A data set object}

\item{mutation_index}{Index of mutations to be removed}
}
\value{
a data set in which the pseudo SNVs are removed
}
\description{
Helper function to remove a set of mutations from a dataset by their index
}
