% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadData.R
\name{remove_pseudo_snv_cna_clusters}
\alias{remove_pseudo_snv_cna_clusters}
\title{Replace the pseudo SNV clusters that represent a CNA event during clustering
by amalgamated assignments for each CNA}
\usage{
remove_pseudo_snv_cna_clusters(dataset)
}
\arguments{
\item{dataset}{A data set object}
}
\value{
a data set in which the pseudo SNVs are removed
}
\description{
Replace the pseudo SNV clusters that represent a CNA event during clustering
by amalgamated assignments for each CNA
}
