% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleMutations.R
\name{unsample_mutations}
\alias{unsample_mutations}
\title{Unsample a sampled dataset and expand clustering results with the mutations that were not used during clustering.
Mutations are assigned to the same cluster as the most similar mutation is assigned to}
\usage{
unsample_mutations(dataset, clustering_result)
}
\arguments{
\item{dataset}{A dataset object in which mutations have been sampled}

\item{clustering_result}{A clustering result object based on the downsampled mutations}
}
\value{
A list containing the unsampled dataset and the clustering object with the unused mutations included
}
\description{
Unsample a sampled dataset and expand clustering results with the mutations that were not used during clustering.
Mutations are assigned to the same cluster as the most similar mutation is assigned to
}
