% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{build_coassignment_prob_matrix_preferences}
\alias{build_coassignment_prob_matrix_preferences}
\title{Helper function that builds a mutation to mutation coassignment probability matrix out of mutation to cluster assignments}
\usage{
build_coassignment_prob_matrix_preferences(GS.data, density, no.muts,
  sampledIters)
}
\arguments{
\item{GS.data}{Main clustering output directly from the chain}

\item{density}{Posterior density through the MCMC output}

\item{no.muts}{The number of mutations in the data set}

\item{sampledIters}{Iterations to consider}
}
\value{
A no.muts x no.muts matrix where each cell contains the fraction of iterations the pair of mutations was assigned to the same cluster
}
\description{
Helper function that builds a mutation to mutation coassignment probability matrix out of mutation to cluster assignments
}
