% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignMutations.R
\name{calc_cluster_conf_intervals}
\alias{calc_cluster_conf_intervals}
\title{Calculate confidence intervals on the cluster location}
\usage{
calc_cluster_conf_intervals(GS.data, mut_assignments, clusterids, no.muts,
  no.timepoints, no.iters, no.iters.burn.in)
}
\arguments{
\item{GS.data}{MCMC output with assignments and cluster locations}

\item{mut_assignments}{Final mutation to cluster assignments}

\item{clusterids}{Clusterids to run through}

\item{no.muts}{The total number of mutations}

\item{no.timepoints}{Total number of samples in this dataset}

\item{no.iters}{Total number of iterations}

\item{no.iters.burn.in}{Number of iterations to use as burn-in}
}
\value{
A data.frame with the confidence intervals for each cluster
}
\description{
Calculate confidence intervals on the cluster location
}
\author{
sd11
}
