% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpclust-package.R
\docType{package}
\name{DPClust}
\alias{DPClust}
\alias{DPClust-package}
\title{Dirichlet Process Clustering of mutations}
\description{
Dirichlet Process Clustering of mutations
}
\author{
Stefan Dentro, \email{sd11@sanger.ac.uk}
}
\keyword{package}
