% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirichletProcessClustering.R
\name{make_sample_params}
\alias{make_sample_params}
\title{Helper function to package sample parameters}
\usage{
make_sample_params(datafiles, cellularity, is.male, samplename, subsamples,
  mutphasingfiles = NULL)
}
\arguments{
\item{datafiles}{Vector of data files to be read in}

\item{cellularity}{Vector with purity for all samples}

\item{is.male}{Boolean whether the sample is male}

\item{samplename}{Donorname, used in plots and to name output files}

\item{subsamples}{Samplenames, used for multi-dimensional clustering to denote individual samples (this information is thought of as a suffix to the samplename, so: samplename=PD4120 and subsamplename=c(a, c)}
}
\value{
A list containing these components
}
\description{
Helper function to package sample parameters
}
\author{
sd11
}
