% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignMutations.R
\name{oneDimensionalClustering}
\alias{oneDimensionalClustering}
\title{Identify clusters and assign mutations for 1-dimensional clustering}
\usage{
oneDimensionalClustering(samplename, subclonal.fraction, GS.data, density,
  no.iters, no.iters.burn.in)
}
\arguments{
\item{samplename}{Sample identifier, used for writing out data to disk}

\item{subclonal.fraction}{Cancer cell fraction estimates for each mutation}

\item{GS.data}{Output from the MCMC chain}

\item{density}{Posterior density estimate of where clusters are located}

\item{no.iters}{Number of total iterations the MCMC chain was run for}

\item{no.iters.burn.in}{Number of iterations to discard as burn in}
}
\value{
Standardised mutation clustering output, including clusters, mutation assignments and likelihoods
}
\description{
Identify clusters and assign mutations for 1-dimensional clustering
}
\author{
dw9, sd11
}
