% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDensities.R
\name{plot1D_2}
\alias{plot1D_2}
\title{ggplot2 based density figure that can show CNA-pseudo-SNVs separately. Currently this figure only can plot the CCF space.}
\usage{
plot1D_2(density, polygon.data, mutationCopyNumber, no.chrs.bearing.mut,
  pngFile = NA, density.from = 0, x.max = NA, y.max = NA,
  y = NULL, N = NULL, samplename = "", CALR = numeric(0),
  cluster.locations = NULL, mutation.assignments = NULL,
  mutationTypes = rep("SNV", length(mutationCopyNumber)),
  font_sizes = list(plot.title = 50, axis.text = 25, axis.title = 35,
  legend.text = 25, legend.title = 25, legend.position = "bottom"),
  cbPalette = c("lightgray", "red", "blue"))
}
\arguments{
\item{density}{Density as returned by Gibbs.subclone.density.est.1d.}

\item{polygon.data}{Polygon data as returned by Gibbs.subclone.density.est.1d.}

\item{mutationCopyNumber}{Mutation copy number matrix to be used to construct the CCF space.}

\item{no.chrs.bearing.mut}{Multiplicity matrix to be used to construct the CCF space.}

\item{pngFile}{Filename of the PNG file to be created. If NA is supplied the figure is returned (Default: NA).}

\item{density.from}{Legacy parameter that is no longer used (Default: 0).}

\item{x.max}{Max value on the x-axis to be plotted (Default: NA, then determined based on the data).}

\item{y.max}{Max value on the y-axis to be plotted (Default: NA, then determined based on the data).}

\item{y}{Legacy parameter that is used in the original figure to be able to plot different spaces. This value should be the number of supporting variant reads. It has no effect here. (Default: NULL).}

\item{N}{Legacy parameter that is used in the original figure to be able to plot different spaces. This value should be the total number of reads. It has no effect here. (Default: NULL).}

\item{samplename}{Name of the sample to be put on top of the figure.}

\item{CALR}{Legacy parameter that is no longer used.}

\item{cluster.locations}{Locations of where clusters were found. A vectical line is plotted for each cluster.}

\item{mutation.assignments}{Vector with hard assignment of mutations to clusters - optional, is used for plotting (Default: NULL)}

\item{mutationTypes}{Vector with mutation types. The background histogram can be coloured by mutation type (Default: SNV)}

\item{font_sizes}{A list that defines the size of the fonts in the title, axis and legends (Default: Best settings for 1500x1000 plot).}

\item{cbPalette}{Colours for mutation types: Grey for first mutation type (SNVs), orange for second (CNAs), as defined in LoadData}
}
\description{
ggplot2 based density figure that can show CNA-pseudo-SNVs separately. Currently this figure only can plot the CCF space.
}
\author{
sd11
}
