% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDensities.R
\name{plotAssignmentTable}
\alias{plotAssignmentTable}
\title{Plot a table with the assignment counts}
\usage{
plotAssignmentTable(cluster_locations, pngFile, cndata = NULL,
  num_samples = 1, indeldata = NULL)
}
\arguments{
\item{cluster_locations}{Cluster table with cluster number, cluster location and number of mutations as columns.}

\item{pngFile}{Output file to save the image.}

\item{cndata}{Optional CNA data table. This must be the table from after assigning CNA events to clusters (Default: NULL).}

\item{num_samples}{Optional parameter representing the number of samples that have been clustered (Default: 1)}

\item{indeldata}{Optional indel data table. This must be the table from after assigning indel events to clusters (Default: NULL).}
}
\description{
Plot a table with the assignment counts
}
\author{
sd11
}
