% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirichletProcessClustering.R
\name{RunDP}
\alias{RunDP}
\title{Main DPClust function that handles the various pipelines}
\usage{
RunDP(analysis_type, run_params, sample_params, advanced_params, outdir,
  cna_params = NULL, mutphasingfiles = NULL)
}
\arguments{
\item{analysis_type}{Type of analysis to run: nd_dp (1d and nd clustering), replot_1d/replot_nd (recreate plots), reassign_muts_1d/reassign_muts_nd (reassign mutations)}

\item{run_params}{List with run parameters (see make_run_params)}

\item{sample_params}{List with sample parameters (see make_sample_params)}

\item{advanced_params}{List with advanced parameters (see make_advanced_params)}

\item{outdir}{Directory where the output will be saved}

\item{cna_params}{List with copy number parameters - currently unsupported (Default: NULL)}

\item{mutphasingfiles}{Mutation phasing files - currently unsupported (Default: NULL)}
}
\description{
Main DPClust function that handles the various pipelines
}
\author{
sd11
}
