% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirichletProcessClustering.R
\name{writeStandardFinalOutput}
\alias{writeStandardFinalOutput}
\title{Function that stores the final output in a unified format on disk}
\usage{
writeStandardFinalOutput(clustering, dataset, most.similar.mut,
  outfiles.prefix, outdir, samplename, subsamplenames, GS.data, density,
  polygon.data, no.iters, no.iters.burn.in, min_muts_cluster,
  min_frac_muts_cluster, assign_sampled_muts = T, write_tree = F,
  generate_cluster_ordering = F, no.samples.cluster.order = 1000)
}
\arguments{
\item{clustering}{A clustering result}

\item{dataset}{The dataset that went into clustering}

\item{most.similar.mut}{Vector containing for each non-sampled mutation its most similar sampled mutation. The non-sampled mutation will be assigned to the same cluster}

\item{outfiles.prefix}{A prefix for the filenames}

\item{outdir}{Output directory where the replot of the 1D method is to be stored if clusters are removed due to being too small}

\item{samplename}{Overall samplename}

\item{subsamplenames}{Samplenames of the different timepoints}

\item{GS.data}{MCMC output}

\item{density}{Posterior density estimate across MCMC iterations}

\item{polygon.data}{1D confidence interval, used for plotting the 1D method (set to NA for multi-D method)}

\item{no.iters}{Number of total iterations}

\item{no.iters.burn.in}{Number of iterations to be used as burn-in}

\item{min_muts_cluster}{The minimum number of mutations required for a cluster to be kept in the final output}

\item{min_frac_muts_cluster}{The minimum fraction of mutations required for a cluster to be kept in the final}

\item{assign_sampled_muts}{Boolean whether to assign the non-sampled mutations (Default: TRUE)}

\item{write_tree}{Boolean whether to write a tree to file. Not all clustering methods return a tree (Default: FALSE)}

\item{generate_cluster_ordering}{Boolean specifying whether a possible cluster ordering should be determined (Default: FALSE)}

\item{no.samples.cluster.order}{Number of mutations to sample (with replacement) to classify pairs of clusters into parent-offspring or siblings (Default: 1000)}
}
\description{
Function that stores the final output in a unified format on disk
}
\author{
sd11
}
